implement on Delete Function

+++ src/main/java/org/app/service/ProductServiceImp.java
@@ -3,10 +3,13 @@ package org.app.service;
 import org.app.model.Product;
 import org.app.repo.ProductRepoImp;
 import org.app.utilies.TableConfig;
+import org.app.utilies.UserInput;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Scanner;
+
 
 public class ProductServiceImp implements ProductService {
     ProductRepoImp productRepoImp = new ProductRepoImp();
@@ -28,8 +31,38 @@ public class ProductServiceImp implements ProductService {
     }
 
     @Override
-    public boolean deleteProduct(int id) {
-        return productRepoImp.deleteProduct(id);
+    public void deleteProduct(int id) {
+        try {
+            Scanner scanner = new Scanner(System.in);
+            Product foundProduct = getProductById(id);
+            String yN;
+
+            TableConfig.getTable(List.of(foundProduct));
+
+            loop:
+            do {
+                System.out.print("Are you sure to delete product id: " + id + "? (Y/N) : ");
+                yN = scanner.nextLine();
+                switch (yN.toUpperCase()) {
+                    case "Y":
+                        System.out.println();
+                        productRepoImp.deleteProduct(id);
+                        System.out.println("Product deleted successfully");
+                        System.out.println("Enter to continue...");
+                        scanner.nextLine();
+                        break loop;
+                    case "N":
+                        System.out.println("Exit");
+                        ;
+                        break loop;
+                    default:
+                        System.out.println("Please input Y or N");
+                }
+            } while (true);
+
+        } catch (NullPointerException e) {
+            throw new NullPointerException("Product not found!");
+        }
     }
 
     @Override
